Adding in Archive Page for Physician directory.
I could not get the FacetWP filters to work with a template on a basic page so I integrated in an archive page with Facet widgets. This is working but in the future, I want to go back and figure out why the basic base loading my wp_query wasn't working.